+++ apps/app/pages/[workspaceSlug]/index.tsx
@@ -49,7 +49,7 @@ const WorkspacePage: NextPage = () => {
       )}
       <div className="p-8">
         <div className="flex flex-col gap-8">
-          <div className="text-brand-muted-1 flex flex-col justify-between gap-x-2 gap-y-6 rounded-lg border border-brand-base bg-brand-base px-8 py-6 md:flex-row md:items-center md:py-3">
+          <div className="text-brand-muted-1 flex flex-col justify-between gap-x-2 gap-y-6 rounded-lg border border-brand-base bg-brand-base px-4 py-6 md:flex-row md:items-center md:py-3">
             <p className="font-semibold">
               Plane is open source, support us by starring us on GitHub.
             </p>
